requirements.txt
django-modeltranslation==0.11